Cleaned up code based on comments on PR
